src/auth-service/requirements.txt
astroid
click
Flask
Flask-MySQLdb
isort
itsdangerous
jedi
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
mysqlclient
parso
platformdirs
PyJWT
pylint
python-dotenv
toml
Werkzeug
wrapt